eproject.el: don't barf on non file backed buffers
While eproject-maybe-turn-on is only called on various mode hooks it
breaks on temporary buffers if they use a programming mode. One such
example is when exporting an org-mode document with embedded source
code.

This stops it throwing an error by skipping the attempt if there is no
associated (buffer-file-name) for the buffer.